Fix EventEmitter memory leak in Edge's getstats implementation
