+++ lib/getstats.js
@@ -1,5 +1,6 @@
 'use strict';
 
+var edgeGetTrackStats = require('./edgestats').edgeGetTrackStats;
 var flatMap = require('./util').flatMap;
 var guessBrowser = require('./util').guessBrowser;
 
@@ -301,7 +302,7 @@ function getTrackStats(peerConnection, track, options) {
     return firefoxGetTrackStats(peerConnection, track);
   }
   if (browser === 'edge' || options.testForEdge) {
-    return edgeGetTrackStats(peerConnection, track);
+    return edgeGetTrackStats(peerConnection, track).then(standardizeEdgeAndFirefoxStats);
   }
   return Promise.reject(new Error('RTCPeerConnection#getStats() not supported'));
 }
@@ -318,43 +319,6 @@ function chromeGetTrackStats(peerConnection, track) {
       resolve(standardizeChromeStats(response, track));
     }, null, reject);
   });
-}
-
-/**
- * Get the standardized statistics for a particular MediaStreamTrack in Edge.
- * @param {RTCPeerConnection} peerConnection
- * @param {MediaStreamTrack} track
- * @returns {Promise.<StandardizedTrackStatsReport>}
- */
-function edgeGetTrackStats(peerConnection, track) {
-  // NOTE(syerrapragada): getStats Promise is rejected if ice state is not completed
-  // https://msdn.microsoft.com/en-us/library/mt599588(v=vs.85).aspx
-  return waitForEdgeIceCompletion(peerConnection).then(function() {
-    return peerConnection.getStats(track);
-  }).then(standardizeEdgeAndFirefoxStats);
-}
-
-/**
- * Waits for ice connection state is completed or failed.
- *
- * @param {RTCPeerConnection} peerConnection
- * @returns {Promise<void>}
- */
-function waitForEdgeIceCompletion(peerConnection) {
-  if (peerConnection.iceConnectionState === 'new'
-    || peerConnection.iceConnectionState === 'checking'
-    || peerConnection.iceConnectionState === 'connected') {
-    return new Promise(function(resolve) {
-      peerConnection.addEventListener('iceconnectionstatechange', function onIceConnectionStateChanged() {
-        if (peerConnection.iceConnectionState === 'completed' || peerConnection.iceConnectionState === 'failed') {
-          peerConnection.removeEventListener('iceconnectionstatechange', onIceConnectionStateChanged);
-          // NOTE(syerrapragada): Without a 250ms delay getStats method returns a promise that is rejected
-          setTimeout(resolve, 250);
-        }
-      });
-    });
-  }
-  return Promise.resolve();
 }
 
 /**
